@@ -2,24 +2,29 @@
   "name": "gameboilerplate",
   "private": true,
   "version": "0.1.0",
+  "packageManager": "npm@10.8.2",
+  "engines": {
+    "node": ">=22.0.0",
+    "npm": ">=10.0.0"
+  },
   "workspaces": [
     "packages/*",
     "tests"
   ],
   "devDependencies": {
     "@types/jest": "^30.0.0",
-    "@typescript-eslint/eslint-plugin": "^7.5.0",
-    "@typescript-eslint/parser": "^7.5.0",
+    "@typescript-eslint/eslint-plugin": "^8.15.0",
+    "@typescript-eslint/parser": "^8.15.0",
     "dotenv": "^17.2.1",
-    "eslint": "^8.56.0",
+    "eslint": "^9.17.0",
     "eslint-config-prettier": "^10.1.8",
-    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-prettier": "^5.2.1",
     "husky": "^9.1.7",
     "lint-staged": "^15.5.2",
-    "prettier": "^3.3.0",
+    "prettier": "^3.4.2",
     "socket.io-client": "^4.8.1",
-    "turbo": "^1.12.0",
-    "typescript": "^5.4.0"
+    "turbo": "^2.3.3",
+    "typescript": "^5.7.2"
   },
   "scripts": {
     "build": "npx turbo run build",
