@@ -2,6 +2,10 @@
   "name": "@gameboilerplate/admin",
   "version": "0.1.0",
   "private": true,
+  "engines": {
+    "node": ">=22.0.0",
+    "npm": ">=10.0.0"
+  },
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
@@ -14,6 +18,7 @@
     "@mui/icons-material": "^7.2.0",
     "@mui/material": "^7.2.0",
     "@mui/x-charts": "^8.9.0",
+    "@vitejs/plugin-react": "^4.7.0",
     "chart.js": "^4.5.0",
     "react": "^18.3.1",
     "react-chartjs-2": "^5.3.0",
@@ -25,8 +30,8 @@
     "@types/react": "^18.3.23",
     "@types/react-dom": "^18.3.7",
     "@types/react-router-dom": "^5.3.3",
-    "typescript": "^5.4.0",
-    "vite": "^7.0.6"
+    "typescript": "^5.7.2",
+    "vite": "^6.0.7"
   },
   "jest": {
     "setupFilesAfterEnv": [
