@@ -2,11 +2,16 @@
   "name": "@gameboilerplate/server",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
+  "engines": {
+    "node": ">=22.0.0",
+    "npm": ">=10.0.0"
+  },
   "scripts": {
-    "dev": "ts-node-dev src/index.ts",
-    "dev:mock": "MOCK_MODE=true ts-node-dev src/index.ts",
+    "dev": "tsx watch src/index.ts",
+    "dev:mock": "MOCK_MODE=true tsx watch src/index.ts",
     "build": "tsc -p tsconfig.json",
     "start": "node dist/index.js",
     "clean": "rm -rf dist",
@@ -17,17 +22,18 @@
     "bcrypt": "^6.0.0",
     "cors": "^2.8.5",
     "dotenv": "^17.2.1",
-    "express": "^4.21.2",
+    "express": "^5.0.1",
     "jsonwebtoken": "^9.0.2",
-    "mongoose": "^8.17.0",
+    "mongoose": "^8.9.3",
     "socket.io": "^4.8.1",
     "zod": "^4.0.14"
   },
   "devDependencies": {
     "@types/cors": "^2.8.19",
-    "@types/express": "^4.17.23",
+    "@types/express": "^5.0.0",
     "@types/jsonwebtoken": "^9.0.10",
     "ts-node-dev": "^2.0.0",
-    "typescript": "^5.4.0"
+    "tsx": "^4.20.3",
+    "typescript": "^5.7.2"
   }
 }
