@@ -2,6 +2,10 @@
   "name": "@gameboilerplate/tests",
   "version": "0.1.0",
   "private": true,
+  "engines": {
+    "node": ">=22.0.0",
+    "npm": ">=10.0.0"
+  },
   "scripts": {
     "test": "jest --config=config/jest.config.js --passWithNoTests",
     "test:unit": "jest --config=config/jest.unit.config.js",
@@ -15,12 +19,12 @@
     "test:shared:coverage": "jest --config=config/jest.unit.config.js --testPathPatterns=shared --coverage"
   },
   "devDependencies": {
-    "@babel/core": "^7.28.0",
-    "@babel/preset-env": "^7.28.0",
+    "@babel/core": "^7.26.0",
+    "@babel/preset-env": "^7.26.0",
     "@playwright/test": "^1.54.1",
     "@testing-library/jest-dom": "^6.6.4",
     "@testing-library/react": "^16.3.0",
-    "@types/jest": "^29.5.11",
+    "@types/jest": "^30.0.0",
     "@types/pngjs": "^6.0.5",
     "babel-jest": "^30.0.5",
     "jest": "^30.0.5",
@@ -28,10 +32,10 @@
     "playwright": "^1.54.1",
     "pngjs": "^7.0.0",
     "supertest": "^7.1.4",
-    "ts-jest": "^29.1.1",
-    "typescript": "^5.4.0"
+    "ts-jest": "^29.2.6",
+    "typescript": "^5.7.2"
   },
   "dependencies": {
-    "node-fetch": "^2.7.0"
+    "node-fetch": "^3.3.2"
   }
 }
