@@ -19,7 +19,7 @@ interface AuthActions {
   clearError: () => void;
 }
 
-const API_BASE_URL = 'http://localhost:3001';
+const API_BASE_URL = 'http://localhost:3000';
 
 export const useAuthStore = create<AuthState & AuthActions>()(
   persist(
