@@ -1,9 +1,9 @@
-export * from './GameEngine';
-export * from './ActionSystem';
-export * from './SystemManager';
-export * from './SceneManager';
-export * from './PluginSystem';
-export * from './types';
+export * from './GameEngine.js';
+export * from './ActionSystem.js';
+export * from './SystemManager.js';
+export * from './SceneManager.js';
+export * from './PluginSystem.js';
+export * from './types.js';
 
 // Example plugins
-export { FarmingPlugin } from './plugins/FarmingPlugin';
+export { FarmingPlugin } from './plugins/FarmingPlugin.js';
